Add Detector

+++ Framework/Core/include/Framework/MessageContext.h
@@ -503,7 +503,7 @@ class MessageContext
             outputsPerChannel[mProxy.getOutputChannelIndex(message->route()).value].AddPart(std::move(part));
           }
         }
-        for (int ci = 0; ci < mProxy.getNumOutputChannels(); ++ci) {
+        for (size_t ci = 0; ci < mProxy.getNumOutputChannels(); ++ci) {
           auto& parts = outputsPerChannel[ci];
           if (parts.Size() == 0) {
             continue;
